requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
google-cloud-storage==2.13.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
aiohttp
aiomysql==0.2.0
passlib==1.7.4
python-jose==3.3.0